feat(#3251): pid

+++ eo-runtime/src/test/java/EOorg/EOeolang/EOsys/EOposixTest.java
@@ -62,9 +62,9 @@ final class EOposixTest {
                     "args",
                     new EOtuple$EOempty()
                 ).take("code")
-            ).asNumber(),
+            ).asNumber().intValue(),
             Matchers.equalTo(
-                (double) Long.parseLong(
+                Integer.parseInt(
                     ManagementFactory.getRuntimeMXBean()
                         .getName().split("@")[0]
                 )
